requirements.txt
docker
PyYAML
slackclient
